requirements.txt
pyyaml
pytest
pytest-asyncio==1.3.0
aiohttp
